+++ coreutils/od.rs
@@ -1034,7 +1034,7 @@ N_SKIP bytes in the combined input, give an error message and return
 nonzero.  When possible, use seek rather than read operations to
 advance IN_STREAM.  */
 unsafe extern "C" fn skip(mut n_skip: off_t) {
-  if n_skip == 0i32 as libc::c_long {
+  if n_skip ==0{
     return;
   }
   while !(*(bb_common_bufsiz1.as_mut_ptr() as *mut globals))
@@ -1062,8 +1062,7 @@ unsafe extern "C" fn skip(mut n_skip: off_t) {
       &mut file_stats,
     ) == 0i32
       && file_stats.st_mode & 0o170000i32 as libc::c_uint == 0o100000i32 as libc::c_uint
-      && file_stats.st_size > 0i32 as libc::c_long
-    {
+      && file_stats.st_size > 0     {
       if file_stats.st_size < n_skip {
         n_skip -= file_stats.st_size
       /* take "check & close / open_next" route */
@@ -1084,7 +1083,7 @@ unsafe extern "C" fn skip(mut n_skip: off_t) {
       let mut buf: [libc::c_char; 1024] = [0; 1024];
       let mut n_bytes_to_read: size_t = 1024i32 as size_t;
       let mut n_bytes_read: size_t = 0;
-      while n_skip > 0i32 as libc::c_long {
+      while n_skip >0{
         if (n_skip as libc::c_ulong) < n_bytes_to_read {
           n_bytes_to_read = n_skip as size_t
         }
@@ -1101,7 +1100,7 @@ unsafe extern "C" fn skip(mut n_skip: off_t) {
         /* EOF on this file or error */
       }
     }
-    if n_skip == 0i32 as libc::c_long {
+    if n_skip ==0{
       return;
     }
     check_and_close();
@@ -1645,7 +1644,7 @@ unsafe extern "C" fn parse_old_offset(
   if !p.is_null() {
     *p.offset(0) = '.' as i32 as libc::c_char
   }
-  return (*offset >= 0i32 as libc::c_long) as libc::c_int;
+  return (*offset >=0) as libc::c_int;
 }
 
 #[no_mangle]
@@ -1844,7 +1843,7 @@ pub unsafe extern "C" fn od_main(
       } else {
         bb_simple_error_msg_and_die(b"too many arguments\x00" as *const u8 as *const libc::c_char);
       }
-      if pseudo_start >= 0i32 as libc::c_long {
+      if pseudo_start >=0{
         if (*(bb_common_bufsiz1.as_mut_ptr() as *mut globals)).format_address
           == Some(format_address_none as unsafe extern "C" fn(_: off_t, _: libc::c_char) -> ())
         {
